redesigned the invite review screen
