+++ packages/mobile/src/account/InviteReview.tsx
@@ -1,8 +1,9 @@
 import Button, { BtnTypes } from '@celo/react-components/components/Button'
+import HorizontalLine from '@celo/react-components/components/HorizontalLine'
 import ReviewFrame from '@celo/react-components/components/ReviewFrame'
 import ReviewHeader from '@celo/react-components/components/ReviewHeader'
 import colors from '@celo/react-components/styles/colors'
-import { CURRENCY_ENUM } from '@celo/utils/src/currencies'
+import { CURRENCIES, CURRENCY_ENUM } from '@celo/utils/src/currencies'
 import { StackScreenProps } from '@react-navigation/stack'
 import BigNumber from 'bignumber.js'
 import * as React from 'react'
@@ -13,7 +14,11 @@ import { connect } from 'react-redux'
 import { hideAlert, showError } from 'src/alert/actions'
 import CeloAnalytics from 'src/analytics/CeloAnalytics'
 import { CustomEventNames } from 'src/analytics/constants'
-import { TokenTransactionType } from 'src/apollo/types'
+import Avatar from 'src/components/Avatar'
+import CurrencyDisplay, { DisplayType, FormatType } from 'src/components/CurrencyDisplay'
+import FeeIcon from 'src/components/FeeIcon'
+import LineItemRow from 'src/components/LineItemRow.v2'
+import TotalLineItem from 'src/components/TotalLineItem.v2'
 import { FeeType } from 'src/fees/actions'
 import CalculateFee from 'src/fees/CalculateFee'
 import { getFeeDollars } from 'src/fees/selectors'
@@ -27,7 +32,6 @@ import { navigateBack } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { StackParamList } from 'src/navigator/types'
 import { RootState } from 'src/redux/reducers'
-import TransferReviewCard from 'src/send/TransferReviewCard'
 import { fetchDollarBalance } from 'src/stableToken/actions'
 import { currentAccountSelector } from 'src/web3/selectors'
 
@@ -164,7 +168,7 @@ export class InviteReview extends React.Component<Props, State> {
           disabled={inviteInProgress}
         />
         {inviteInProgress && (
-          <View style={style.loadingIcon}>
+          <View style={styles.loadingIcon}>
             <ActivityIndicator size="large" color={colors.celoGreen} />
           </View>
         )}
@@ -173,32 +177,76 @@ export class InviteReview extends React.Component<Props, State> {
   }
 
   render() {
-    const { account } = this.props
+    const { account, t } = this.props
     if (!account) {
       throw Error('Account is required')
     }
 
     const recipient = this.getRecipient()
-    const amount = getInvitationVerificationFeeInDollars()
+    const inviteFee = getInvitationVerificationFeeInDollars()
+
+    const inviteFeeAmount = {
+      value: inviteFee,
+      currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code,
+    }
 
     return (
-      <SafeAreaView style={style.container}>
+      <SafeAreaView style={styles.container}>
         <ReviewFrame HeaderComponent={this.renderHeader} FooterComponent={this.renderFooter}>
-          <CalculateFee feeType={FeeType.INVITE} account={account} amount={amount} comment="">
-            {(asyncFee) => (
-              <TransferReviewCard
-                recipient={recipient}
-                type={TokenTransactionType.InviteSent}
-                address={recipient.address}
-                // It's only an invite without additional funds
-                value={new BigNumber(0)}
-                e164PhoneNumber={recipient.e164PhoneNumber}
-                currency={CURRENCY_ENUM.DOLLAR}
-                fee={getFeeDollars(asyncFee.result)}
-                isLoadingFee={asyncFee.loading}
-                feeError={asyncFee.error}
-              />
-            )}
+          <CalculateFee
+            feeType={FeeType.INVITE}
+            account={account}
+            amount={new BigNumber(0)}
+            comment=""
+          >
+            {(asyncFee) => {
+              const fee = getFeeDollars(asyncFee.result)
+              const securityFee = fee ? fee.minus(inviteFee) : fee
+              const securityFeeAmount = securityFee && {
+                value: securityFee,
+                currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code,
+              }
+
+              const totalAmount = {
+                value: fee || 0,
+                currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code,
+              }
+
+              return (
+                <View>
+                  <Avatar
+                    recipient={recipient}
+                    address={recipient.address}
+                    e164Number={recipient.e164PhoneNumber}
+                  />
+                  <CurrencyDisplay
+                    type={DisplayType.Big}
+                    style={styles.amount}
+                    amount={inviteFeeAmount}
+                  />
+                  <View style={styles.bottomContainer}>
+                    <HorizontalLine />
+                    <LineItemRow
+                      title={t('inviteFee')}
+                      amount={<CurrencyDisplay amount={inviteFeeAmount} />}
+                    />
+                    <LineItemRow
+                      title={t('sendFlow7:securityFee')}
+                      titleIcon={<FeeIcon />}
+                      amount={
+                        securityFeeAmount && (
+                          <CurrencyDisplay amount={securityFeeAmount} formatType={FormatType.Fee} />
+                        )
+                      }
+                      isLoading={asyncFee.loading}
+                      hasError={!!asyncFee.error}
+                    />
+                    <HorizontalLine />
+                    <TotalLineItem amount={totalAmount} />
+                  </View>
+                </View>
+              )
+            }}
           </CalculateFee>
         </ReviewFrame>
       </SafeAreaView>
@@ -206,9 +254,22 @@ export class InviteReview extends React.Component<Props, State> {
   }
 }
 
-const style = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
     flex: 1,
+    justifyContent: 'flex-start',
+    paddingBottom: 25,
+  },
+  editContainer: {
+    flexDirection: 'row',
+  },
+  bottomContainer: {
+    marginTop: 5,
+    flexDirection: 'column',
+    alignItems: 'stretch',
+  },
+  amount: {
+    marginTop: 15,
   },
   loadingIcon: {
     position: 'absolute',
